Update Fibonacci.py
Changed the Fibonacci code to prompt the user for the number of terms and print that many numbers from the sequence, rather than just returning the nth Fibonacci number. This makes the script interactive by displaying the entire Fibonacci sequence up to the user’s chosen length.

Please consider this code.